update handling of company address options
