+++ src/composables/useSelectedOrganization.ts
@@ -7,7 +7,6 @@ import { useRegisterChallengeStore } from '../stores/registerChallenge';
 import type { Organization } from '../components/types/Organization';
 import type {
   FormCompanyAddressFields,
-  FormOption,
   FormSelectTableOption,
 } from 'src/components/types/Form';
 
@@ -64,61 +63,9 @@ export const useSelectedOrganization = (organizations: Organization[]) => {
     set: (value: number | null) => store.setSubsidiaryId(value),
   });
 
-  /**
-   * Computes the address options for the address select
-   * based on the selected organizationId.
-   * @returns {FormSelectTableOption[]} - The address options or empty array.
-   */
-  const addressOptions = computed<FormOption[]>(() => {
-    if (!organizationId.value) return [];
-
-    const selectedOrganization = organizations.find(
-      (organization) => organization.id === organizationId.value,
-    );
-    if (!selectedOrganization) return [];
-
-    const addressOptions: FormOption[] = [] as FormOption[];
-
-    // add subsidiaries addresses to the options
-    selectedOrganization.subsidiaries.forEach((subsidiary) => {
-      if (subsidiary.address) {
-        addressOptions.push({
-          label: getAddressString(subsidiary.address),
-          value: subsidiary.address,
-        });
-      }
-    });
-
-    return addressOptions;
-  });
-
-  /**
-   * Get a formatted address string from the provided address object.
-   * @param {FormCompanyAddressFields | undefined} address - The address object.
-   * @returns {string} - Formatted string representation of the address or
-   * empty string.
-   */
-  function getAddressString(
-    address: FormCompanyAddressFields | undefined,
-  ): string {
-    if (!address) return '';
-
-    const parts = [
-      address.street,
-      address.houseNumber,
-      address.city,
-      address.zip,
-      address.cityChallenge,
-      address.department,
-    ].filter(Boolean);
-
-    return parts.join(', ');
-  }
-
   return {
     selectedAddress,
     subsidiaryId,
-    addressOptions,
     organizationId,
     organizationOptions,
   };
